# api-design/autogenerated/openapi.yaml
openapi: 3.0.0
info:
  title: TodoDoo API
  version: 0.0.0
...
paths:
...
  /api/funs:
    get:
      operationId: Funs_list
      description: List all funs for the authenticated user
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fun'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
    post:
      operationId: Funs_create
      description: Create a new fun
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fun'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFunRequest'
  /api/funs/{id}:
    get:
      operationId: Funs_get
      description: Get a specific fun by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fun'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
    patch:
      operationId: Funs_update
      description: Update a fun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fun'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFunRequest'
    delete:
      operationId: Funs_delete
      description: Delete a fun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
...
  /api/todos:
    get:
      operationId: Todos_list
      description: List all todos for the authenticated user
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
    post:
      operationId: Todos_create
      description: Create a new todo
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
...
  /api/todos/{id}:
    get:
      operationId: Todos_get
      description: Get a specific todo by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
    patch:
      operationId: Todos_update
      description: Update a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
    delete:
      operationId: Todos_delete
      description: Delete a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
...
  /api/weighted-random-funs:
    get:
      operationId: WeightedRandomFuns_list
      description: Get weighted random funs for the authenticated user
      parameters:
        - name: 'n'
          in: query
          required: true
          schema:
            type: integer
            format: int32
          explode: false
        - name: excludeIds
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fun'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Funs
  /api/weighted-random-todos:
    get:
      operationId: WeightedRandomTodos_list
      description: Get weighted random todos for the authenticated user
      parameters:
        - name: 'n'
          in: query
          required: true
          schema:
            type: integer
            format: int32
          explode: false
        - name: maxTimeEstimate
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TimeEstimate'
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
security:
  - BearerAuth: []
components:
  schemas:
    Annoyingness:
      type: integer
      format: int32
      minimum: 0
      maximum: 4
      description: Annoyingness level (0-4)
    CreateFunRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: The fun title
        position:
          allOf:
            - $ref: '#/components/schemas/TodoPosition'
          description: Where to insert the fun (defaults to top)
      description: Request body for creating a new fun
...
    CreateTodoRequest:
      type: object
      required:
        - task
      properties:
        task:
          type: string
          description: The task description
        details:
          type: object
          description: Additional details stored as JSON
        status:
          allOf:
            - $ref: '#/components/schemas/TodoStatus'
          description: Current status of the todo (defaults to 'todo')
        position:
          allOf:
            - $ref: '#/components/schemas/TodoPosition'
          description: Where to insert the todo (defaults to top)
        importance:
          allOf:
            - $ref: '#/components/schemas/Importance'
          description: Importance level (0-5)
        time_estimate:
          allOf:
            - $ref: '#/components/schemas/TimeEstimate'
          description: Estimated time to complete
        annoyingness:
          allOf:
            - $ref: '#/components/schemas/Annoyingness'
          description: Annoyingness level (0-4)
        committed:
          type: boolean
          description: Whether the todo is committed (defaults to false)
        atomic:
          type: boolean
          description: Whether the todo is atomic (defaults to false)
        infinitely_divisible:
          type: boolean
          description: Whether the todo is infinitely divisible (defaults to false)
      description: Request body for creating a new todo
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      description: Error response
...
    Fun:
      type: object
      required:
        - id
        - title
        - user_id
        - order
        - date_created
        - date_updated
      properties:
        id:
          type: string
          description: Unique identifier for the fun
        title:
          type: string
          description: The fun title
        user_id:
          type: string
          description: ID of the user who owns this fun
        order:
          type: string
          description: Order/position of the fun in the list
        date_created:
          type: string
          format: date-time
          description: When the fun was created
        date_updated:
          type: string
          format: date-time
          description: When the fun was last updated
      description: A fun item
...
    Importance:
      type: integer
      format: int32
      minimum: 0
      maximum: 5
      description: Importance level (0-5)
...
    PositionAfter:
      type: object
      required:
        - type
        - id
      properties:
        type:
          type: string
          enum:
            - after
        id:
          type: string
          description: ID of the todo to insert after
      description: Insert after a specific todo
    PositionBottom:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - bottom
      description: Insert at the bottom of the list
    PositionTop:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - top
      description: Insert at the top of the list
...
    TimeEstimate:
      type: string
      enum:
        - 1-min
        - 5-mins
        - 15-mins
        - 30-mins
        - 1-hour
        - 2-hours
        - half-day
        - one-day
        - project
      description: Time estimate for completing a todo
    Todo:
      type: object
      required:
        - id
        - task
        - status
        - date_created
        - date_updated
        - user_id
        - order
        - annoyingness
        - committed
        - elapsedTimeSeconds
        - atomic
        - infinitely_divisible
      properties:
        id:
          type: string
          description: Unique identifier for the todo
        task:
          type: string
          description: The task description
        details:
          type: object
          description: Additional details stored as JSON
        status:
          allOf:
            - $ref: '#/components/schemas/TodoStatus'
          description: Current status of the todo
        date_created:
          type: string
          format: date-time
          description: When the todo was created
        date_updated:
          type: string
          format: date-time
          description: When the todo was last updated
        user_id:
          type: string
          description: ID of the user who owns this todo
        order:
          type: string
          description: Order/position of the todo in the list
        priority:
          type: integer
          format: int32
          description: Integer priority position (0-based index in the ordered list)
        desirability:
          type: number
          format: float
          description: Calculated desirability score for weighted random selection
        importance:
          allOf:
            - $ref: '#/components/schemas/Importance'
          description: Importance level (0-5)
        time_estimate:
          allOf:
            - $ref: '#/components/schemas/TimeEstimate'
          description: Estimated time to complete
        annoyingness:
          allOf:
            - $ref: '#/components/schemas/Annoyingness'
          description: Annoyingness level (0-4)
        committed:
          type: boolean
          description: Whether the todo is committed
        elapsedTimeSeconds:
          type: integer
          format: int32
          description: Elapsed time in seconds for timer
        atomic:
          type: boolean
          description: Whether the todo is atomic (cannot be broken down)
        infinitely_divisible:
          type: boolean
          description: Whether the todo is infinitely divisible
      description: A todo item
    TodoPosition:
      anyOf:
        - $ref: '#/components/schemas/PositionTop'
        - $ref: '#/components/schemas/PositionBottom'
        - $ref: '#/components/schemas/PositionAfter'
      description: Position for inserting a new todo
    TodoStatus:
      type: string
      enum:
        - todo
        - complete
        - archived
        - waiting
      description: Status of a todo item
...
    UpdateFunRequest:
      type: object
      properties:
        title:
          type: string
          description: The fun title
      description: Request body for updating a fun
...
    UpdateTodoRequest:
      type: object
      properties:
        task:
          type: string
          description: The task description
        details:
          type: object
          description: Additional details stored as JSON
        status:
          allOf:
            - $ref: '#/components/schemas/TodoStatus'
          description: Current status of the todo
        importance:
          allOf:
            - $ref: '#/components/schemas/Importance'
          description: Importance level (0-5)
        time_estimate:
          allOf:
            - $ref: '#/components/schemas/TimeEstimate'
          description: Estimated time to complete
        annoyingness:
          allOf:
            - $ref: '#/components/schemas/Annoyingness'
          description: Annoyingness level (0-4)
        committed:
          type: boolean
          description: Whether the todo is committed
        atomic:
          type: boolean
          description: Whether the todo is atomic
        infinitely_divisible:
          type: boolean
          description: Whether the todo is infinitely divisible
      description: Request body for updating a todo
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://localhost:8000
    description: Development server
    variables: {}